extract month-selector component
