+++ app/components/month-selector.hbs
@@ -1,0 +1,59 @@
+<div class="box-border flex h-full items-center justify-between gap-4 px-6">
+  <LinkTo
+    @route="year.month"
+    @models={{this.previousMonthRouteModel}}
+    class="rounded p-2 text-red-700 opacity-0 transition-opacity duration-150 ease-in-out hover:bg-gray-100 group-hover:opacity-100"
+  >
+    {{svg-jar "chevron-left" class="size-5"}}
+  </LinkTo>
+  <div class="h-full w-52">
+    <BasicDropdown @horizontalPosition="center" as |dd|>
+      <dd.Trigger
+        class="flex h-full items-center justify-center gap-2 text-xl font-medium text-red-700"
+      >
+        {{this.monthYearLabel}}
+        <div
+          class="opacity-0 transition-opacity duration-150 ease-in-out group-hover:opacity-100"
+        >
+          {{svg-jar "chevron-down" class="size-4"}}
+        </div>
+      </dd.Trigger>
+      <dd.Content>
+        <div class="w-48 rounded-lg border bg-white p-2 shadow">
+          <YearSelector
+            @year={{this.year}}
+            @onDecrementYear={{this.decrementYear}}
+            @onIncrementYear={{this.incrementYear}}
+          />
+          <ul>
+            {{#each this.months as |month|}}
+              <a href="#" {{on "click" (fn this.closeDropdown dd)}}>
+                <LinkTo
+                  @route="year.month"
+                  @models={{array this.year month.humanNumber}}
+                  class="grow text-sm tracking-wide"
+                >
+                  <li
+                    class="flex items-center justify-between rounded p-2 font-light hover:bg-gray-100"
+                  >
+                    {{month.label}}
+                    <span>
+                      <TimesheetStatusIcon @timesheet={{month.timesheet}} />
+                    </span>
+                  </li>
+                </LinkTo>
+              </a>
+            {{/each}}
+          </ul>
+        </div>
+      </dd.Content>
+    </BasicDropdown>
+  </div>
+  <LinkTo
+    @route="year.month"
+    @models={{this.nextMonthRouteModel}}
+    class="rounded p-2 text-red-700 opacity-0 transition-opacity duration-150 ease-in-out hover:bg-gray-100 group-hover:opacity-100"
+  >
+    {{svg-jar "chevron-right" class="size-5"}}
+  </LinkTo>
+</div>
